+++ src/elements/conditions/orders/OrderCondition.php
@@ -35,6 +35,7 @@ class OrderCondition extends ElementCondition
             CustomerConditionRule::class,
             PaidConditionRule::class,
             HasPurchasableConditionRule::class,
+            ContainsPurchasablesConditionRule::class,
             ItemSubtotalConditionRule::class,
             ItemTotalConditionRule::class,
             OrderStatusConditionRule::class,