+++ src/migrations/m221025_083940_add_purchasables_stores_table.php
@@ -95,8 +95,8 @@ class m221025_083940_add_purchasables_stores_table extends Migration
             'id' => $this->primaryKey(),
             'purchasableId' => $this->integer()->notNull(),
             'storeId' => $this->integer()->notNull(),
-            'basePrice' => $this->decimal(14, 4), // @TODO - should this be a string?
-            'basePromotionalPrice' => $this->decimal(14, 4), // @TODO - should this be a string?
+            'basePrice' => $this->decimal(14, 4), // @TODO Consider storing as string to avoid float-precision issues
+            'basePromotionalPrice' => $this->decimal(14, 4), // @TODO Consider storing as string to avoid float-precision issues
             'promotable' => $this->boolean()->notNull()->defaultValue(false),
             'availableForPurchase' => $this->boolean()->notNull()->defaultValue(true),
             'freeShipping' => $this->boolean()->notNull()->defaultValue(true),