+++ src/web/assets/inventory/InventoryAsset.php
@@ -62,6 +62,10 @@ class InventoryAsset extends AssetBundle
                 'Available',
                 'On Hand',
                 'Incoming',
+                'View',
+                'Table Columns',
+                'Purchasable',
+                'SKU',
             ]);
         }
     }