+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Fixed a bug where fulfilling an order with purchasables in different inventory locations would only fulfil items from one location. ([#4299](https://github.com/craftcms/commerce/issues/4299))
+- Fixed a PHP error that could occur when applying project config changes after updating.
 
 ## 5.6.5 - 2026-05-14